filepreviewform is rendering

+++ financial_client/src/components/services/fileService.jsx
@@ -1,12 +1,17 @@
 const FILE_URL = 'http://localhost:3000/api/files'; 
 
 const FileService = {
+    
     addFile: async (files) => {
-        const formData = new FormData();
-        for (const file of files) {
-            formData.append('files', file);
+        if (!Array.isArray(files)) {
+            console.error('Error: files is not an iterable');
+            return [];
         }
-
+        const formData = new FormData();
+        files.forEach(file => {
+            formData.append('files', file);
+        });
+    
         try {
             const response = await fetch(FILE_URL, {
                 method: 'POST',
@@ -17,17 +22,20 @@ const FileService = {
                 const errorMessage = errorBody.error || 'Network response was not ok';
                 throw new Error(errorMessage);
             }
-            return await response.json();
+            return [await response.json()];
         } catch (error) {
             console.error('Error importing files:', error);
-            throw new Error(error.message); 
+            return [];
         }
     },
-
-
     deleteFile: async (hash) => {
+        if (!hash) {
+            console.error('Error: file hash is undefined');
+            return;
+        }
+
         try {
-            const response = await fetch(`${FILE_URL}/delete/${hash}`, {
+            const response = await fetch(`${FILE_URL}/delete/filename/${hash}`, {
                 method: 'DELETE'
             });
             if (!response.ok) {
@@ -40,6 +48,24 @@ const FileService = {
         }
     },
 
+    getFileName: async (hash) => {
+        if (!hash) {
+            console.error('Error: file hash is undefined');
+            return;
+        }
+
+        try {
+            const response = await fetch(`${FILE_URL}/filename/${hash}`);
+            if (!response.ok) {
+                throw new Error('Network response was not ok');
+            }
+            const data = await response.json();
+            return data.filename; 
+        } catch (error) {
+            console.error('Error getting filename:', error);
+            throw error;
+        }
+    }
 };
 
 export default FileService;
